% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_line.R
\name{assert_poll_data_in_time_line_using_time_scale}
\alias{assert_poll_data_in_time_line_using_time_scale}
\alias{assert_poll_data_in_time_line_using_time_range}
\title{Assert that the poll_data object dates are available
within the [time_line] using time_scale}
\usage{
assert_poll_data_in_time_line_using_time_scale(pd, ts)

assert_poll_data_in_time_line_using_time_range(pd, ts, tr)
}
\arguments{
\item{pd}{[polls_data] to check if part of [time_line]}

\item{ts}{a time_scale to use}

\item{tr}{a [time_range]}
}
\description{
The assertion creates a [time_line] using the [time_scale]
}
\seealso{
[time_range()]
}
