% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_data.R, R/stan_polls_data_model.R
\name{assert_stan_data_model}
\alias{assert_stan_data_model}
\alias{assert_stan_data_model.model2}
\alias{assert_stan_data_model.model3}
\alias{assert_stan_data_model.model5}
\alias{assert_stan_data_model.model6}
\alias{assert_stan_data_model.model6b}
\alias{assert_stan_data_model.model8a}
\alias{assert_stan_data_model.model8a1}
\alias{assert_stan_data_model.model8a3}
\alias{assert_stan_data_model.model8c}
\alias{assert_stan_data_model.model8b}
\alias{assert_stan_data_model.model8d}
\alias{assert_stan_data_model.model8e}
\alias{assert_stan_data_model.model8f}
\alias{assert_stan_data_model.model8g}
\alias{assert_stan_data_model.model8i}
\alias{assert_stan_data_model.model8k}
\alias{assert_stan_data_model.model8l}
\alias{assert_stan_data_model.model8m}
\alias{assert_stan_data_model.model4}
\alias{assert_stan_data_model.model9}
\title{Check Stan Data for specific model}
\usage{
assert_stan_data_model(x)

\method{assert_stan_data_model}{model2}(x)

\method{assert_stan_data_model}{model3}(x)

\method{assert_stan_data_model}{model5}(x)

\method{assert_stan_data_model}{model6}(x)

\method{assert_stan_data_model}{model6b}(x)

\method{assert_stan_data_model}{model8a}(x)

\method{assert_stan_data_model}{model8a1}(x)

\method{assert_stan_data_model}{model8a3}(x)

\method{assert_stan_data_model}{model8c}(x)

\method{assert_stan_data_model}{model8b}(x)

\method{assert_stan_data_model}{model8d}(x)

\method{assert_stan_data_model}{model8e}(x)

\method{assert_stan_data_model}{model8f}(x)

\method{assert_stan_data_model}{model8g}(x)

\method{assert_stan_data_model}{model8i}(x)

\method{assert_stan_data_model}{model8k}(x)

\method{assert_stan_data_model}{model8l}(x)

\method{assert_stan_data_model}{model8m}(x)

\method{assert_stan_data_model}{model2}(x)

\method{assert_stan_data_model}{model5}(x)

\method{assert_stan_data_model}{model3}(x)

\method{assert_stan_data_model}{model4}(x)

\method{assert_stan_data_model}{model9}(x)
}
\arguments{
\item{x}{a [poll_of_polls] object}
}
\description{
Check Stan Data for specific model
}
