% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_compute_diagnostics}
\alias{backend_compute_diagnostics}
\title{Compute diagnostics for a backend fit object}
\usage{
backend_compute_diagnostics(backend, fit)
}
\description{
Compute diagnostics for a backend fit object
}
\keyword{internal}
