% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_draws_array}
\alias{backend_draws_array}
\title{Backend posterior draws as a draws_array}
\usage{
backend_draws_array(backend, fit, variables = NULL, inc_warmup = FALSE)
}
\description{
Backend posterior draws as a draws_array
}
\keyword{internal}
