% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_get_num_upars}
\alias{backend_get_num_upars}
\title{Backend number of unconstrained parameters}
\usage{
backend_get_num_upars(backend, fit, ...)
}
\description{
Backend number of unconstrained parameters
}
\keyword{internal}
