% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_get_sampler_params}
\alias{backend_get_sampler_params}
\title{Backend sampler diagnostics}
\usage{
backend_get_sampler_params(backend, fit, inc_warmup = FALSE, ...)
}
\description{
Backend sampler diagnostics
}
\keyword{internal}
