% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_sample}
\alias{backend_sample}
\title{Run a Stan fit using the selected backend}
\usage{
backend_sample(backend, stan_arguments)
}
\description{
Run a Stan fit using the selected backend
}
\keyword{internal}
