% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_constrained_pars_to_unconstrained_pars_map.R
\name{create_constrained_names_to_unconstrained_idx_map}
\alias{create_constrained_names_to_unconstrained_idx_map}
\title{Create a map between constrained parameter names and unconstrained parameter indecies}
\usage{
create_constrained_names_to_unconstrained_idx_map(x, return_first = FALSE)
}
\arguments{
\item{x}{a poll of polls object}

\item{return_first}{return the first character element (the unconstrained,
non-transformed parameter)}
}
\description{
Go through all unconstrained indecies and return all parameters
that changes due to the constrained parameter.

Note that in a state space model a unconstrated parameter can change
a whole series due to the model formulation.
}
