% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_constrained_pars_to_unconstrained_pars_map.R
\name{create_unpar_one_to_one_map}
\alias{create_unpar_one_to_one_map}
\title{Semi-automatic construction of parameters for specific models}
\usage{
create_unpar_one_to_one_map(x)
}
\arguments{
\item{x}{a poll of polls object}
}
\description{
Semi-automatic construction of parameters for specific models
}
\details{
Create the names of the unconstrained parameters from specific models
}
