% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{curate_swedish_polls}
\alias{curate_swedish_polls}
\title{Curate the Swedish polls dataset}
\usage{
curate_swedish_polls(
  x,
  parties = c("M", "L", "C", "KD", "S", "V", "MP", "SD", "FI")
)
}
\arguments{
\item{x}{a SwedishPolls data.frame}

\item{parties}{parties to curate}
}
\description{
Curate the Swedish polls dataset
}
