% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls.R
\name{extract_poll_of_polls_input_arguments}
\alias{extract_poll_of_polls_input_arguments}
\title{Extract all pop arguments for poll_of_polls() from a pop object}
\usage{
extract_poll_of_polls_input_arguments(x)
}
\arguments{
\item{x}{A pop object}
}
\description{
This function takes a pop object and extracts the relevant arguments for the poll_of_polls() function.
It ensures that the pop object is valid and then retrieves the necessary information to be used as input for poll_of_polls().
}
