% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls_extract.R
\name{extract_pop_empirical_posterior_mean_x_change}
\alias{extract_pop_empirical_posterior_mean_x_change}
\title{Extract mean x change from a poll_of_polls object}
\usage{
extract_pop_empirical_posterior_mean_x_change(pop, type)
}
\arguments{
\item{pop}{a poll of polls object}

\item{type}{type of}
}
\description{
Extract mean x change from a poll_of_polls object
}
\details{
Extracts the difference in the latent state between time points,
either as the [diff]erence x[t]-x[t-1] or the [ratio] x[t]/x[t-1].
Here t=1 is set to NA.
}
