% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls.R
\name{format_election_period_to_string}
\alias{format_election_period_to_string}
\title{Format election period lists to string for printout}
\usage{
format_election_period_to_string(x, period_marker = "--", collapse = NULL)
}
\arguments{
\item{x}{a poll of polls object}

\item{period_marker}{character used to combine election period dates to an election period}

\item{collapse}{should the election periods be collapsed to a character of length one (NULL if not),
then the string used to collapse the character vector elements.}
}
\description{
Format election period lists to string for printout
}
