% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls.R
\name{format_obs_x_to_string}
\alias{format_obs_x_to_string}
\title{Format obs_x lists to string for printout}
\usage{
format_obs_x_to_string(x, collapse = NULL)
}
\arguments{
\item{x}{a poll of polls object}

\item{collapse}{should the obs_x characters be collapsed to a character of length one (NULL if not),
then the string used to collapse the character vector elements.}
}
\description{
Format obs_x lists to string for printout
}
