% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{geom_known_state_layers}
\alias{geom_known_state_layers}
\title{Build known-state overlay layers}
\usage{
geom_known_state_layers(data, ...)
}
\arguments{
\item{data}{Normalized overlay data returned by
[known_state_overlay_data()].}

\item{...}{Further arguments passed on to [ggplot2::geom_vline()] and
[ggplot2::geom_point()].}
}
\value{
A list containing the vline and point layers used by
  [geom_known_state()].
}
\description{
Build known-state overlay layers
}
\keyword{internal}
