% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_stan_date}
\alias{get_stan_date}
\title{Extract the date when Stan was run}
\usage{
get_stan_date(object)
}
\arguments{
\item{object}{a [poll_of_polls] object}
}
\description{
Extract the date when Stan was run
}
