% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_line.R
\name{get_time_line}
\alias{get_time_line}
\title{Get time line from polls or time range}
\usage{
get_time_line(x, time_range, time_scale)
}
\arguments{
\item{x}{a [polls_data] object}

\item{time_range}{a time range object}

\item{time_scale}{a time_scale}
}
\description{
Get time line from polls or time range
}
