% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_line.R
\name{get_time_points_from_time_line}
\alias{get_time_points_from_time_line}
\alias{get_total_time_points_from_time_line}
\title{Get time points from a time line object}
\usage{
get_time_points_from_time_line(dates, tl, return_variable = "time_line_t")

get_total_time_points_from_time_line(tl)
}
\arguments{
\item{dates}{a vector with dates to convert to time points}

\item{tl}{a [time_line] object}

\item{return_variable}{the daily variable (time point type) to return}
}
\description{
Look up one index value from the [daily] table of a [time_line] object
for each date in [dates]. This is typically used to convert calendar
dates to latent time-point indices such as [time_line_t].
}
