% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls.R
\name{house_levels}
\alias{house_levels}
\alias{party_levels}
\title{Get the levels/houses from a [poll_of_polls] object}
\usage{
house_levels(x)

party_levels(x)
}
\arguments{
\item{x}{a [poll_of_polls] object}
}
\description{
The house levels are extracted in the same order
as h_i in the stan data object.
}
