% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_state.R
\name{latent_state_FUN}
\alias{latent_state_FUN}
\alias{latent_state_FUN.poll_of_polls}
\alias{latent_state_FUN.latent_state}
\title{Compute an arbitrary function of latent states}
\usage{
latent_state_FUN(x, FUN, var_name)

\method{latent_state_FUN}{poll_of_polls}(x, FUN, var_name)

\method{latent_state_FUN}{latent_state}(x, FUN, var_name)
}
\arguments{
\item{x}{a latent state}

\item{FUN}{an arbitrary function}

\item{var_name}{the variable name to use}
}
\description{
Compute an arbitrary function of latent states
}
