% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls.R
\name{parameter_names}
\alias{parameter_names}
\alias{extract_parameter_names}
\alias{parameter_block_names}
\alias{parameters_counts}
\alias{get_parameter_names}
\title{Extract the parameter names and parameter counts from a poll_of_polls object}
\usage{
parameter_names(x, rm_idx = FALSE)

extract_parameter_names(x, rm_idx = FALSE)

parameter_block_names(x)

parameters_counts(x)

get_parameter_names(x, rm_idx = FALSE)
}
\arguments{
\item{x}{a [poll_of_polls] object}

\item{rm_idx}{remove parameter indecies (e.g. [1], [1,1], [1,1,1]) from the parameter names. Default is FALSE.}
}
\description{
Extract the parameter names and parameter counts from a poll_of_polls object
}
