% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_data.R
\name{parse_obs_x}
\alias{parse_obs_x}
\title{Parse a data frame of observed x (t distributed)}
\usage{
parse_obs_x(hyper_parameters, time_line, y_name)
}
\arguments{
\item{hyper_parameters}{a list that contains the slot obs_x with the dataframe}

\item{time_line}{the time_line to use to parse dates to time points}

\item{y_name}{the different values used to parse obs_x$y to integers}
}
\description{
Parse a data frame of observed x (t distributed)
}
