% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_beta_mu_by_s}
\alias{plot_beta_mu_by_s}
\alias{plot_beta_sigma_by_s}
\alias{plot_beta_mu_by_house}
\alias{plot_beta_mu_by_party}
\alias{plot_beta_mu_by_party_industry_bias_correction}
\alias{plot_beta_sigma_by_house}
\title{Plot beta_mu and beta_sigma by s}
\usage{
plot_beta_mu_by_s(x, y, house)

plot_beta_sigma_by_s(x, house)

plot_beta_mu_by_house(x, y, s, houses = NULL)

plot_beta_mu_by_party(x, y, s, house)

plot_beta_mu_by_party_industry_bias_correction(x, y, s, house)

plot_beta_sigma_by_house(x, s, houses = NULL)
}
\arguments{
\item{x}{a [poll_of_polls] object}

\item{y}{party to plot}

\item{house}{house to plot}

\item{s}{time point to plot}

\item{houses}{houses to plot}
}
\description{
Plot beta_mu and beta_sigma by s
}
