% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_kappa_by_party}
\alias{plot_kappa_by_party}
\alias{plot_kappa_by_s}
\title{Plot kappa}
\usage{
plot_kappa_by_party(x, s, y = NULL)

plot_kappa_by_s(x, y, s = NULL)
}
\arguments{
\item{x}{a [poll_of_polls] object}

\item{s}{time point to plot (NULL is all)}

\item{y}{party to plot (NULL is all)}
}
\description{
Plot kappa
}
