% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pop_draws_array}
\alias{pop_draws_array}
\title{Return posterior draws as a draws_array}
\usage{
pop_draws_array(object, variables = NULL, inc_warmup = FALSE)
}
\description{
Return posterior draws as a draws_array
}
\keyword{internal}
