% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refit_poll_of_polls.R
\name{refit_translate_sample_arguments}
\alias{refit_translate_sample_arguments}
\title{Translate stored sample arguments between backends}
\usage{
refit_translate_sample_arguments(sample_args, from_backend, to_backend)
}
\description{
Translate stored sample arguments between backends
}
\keyword{internal}
