% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs_package.R
\docType{package}
\name{stanpop}
\alias{stanpop-package}
\alias{stanpop}
\title{stanpop: Package to run Stan for poll of polls models.}
\description{
A package to run poll of polls models and Bayesian election prediction models
}
\author{
\strong{Maintainer}: Mans Magnusson \email{mons.magnusson@gmail.com}

}
