% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_stan_basic_on_local}
\alias{test_stan_basic_on_local}
\alias{set_test_stan_basic_on_local}
\alias{test_stan_full_on_local}
\alias{set_test_stan_full_on_local}
\alias{set_on_github_actions}
\title{Run different test suites}
\usage{
test_stan_basic_on_local()

set_test_stan_basic_on_local(x)

test_stan_full_on_local()

set_test_stan_full_on_local(x)

set_on_github_actions(x)
}
\arguments{
\item{x}{value to set it to}
}
\description{
These functions are used to test Stan models locally. As default,
no stan models are tested locally.

To test a basic set of Stan models, those that has been decided to be used,
run \code{test_stan_basic_on_local(TRUE)}.

To run all tests (the whole test suite),
run \code{test_stan_full_on_local(TRUE)} OR run the test suite in the
\code{test} git branch.

To not any longer run a more extensive test suite, run
\code{test_stan_basic_on_local(FALSE)} or \code{test_stan_full_on_local(FALSE)}.

Note! If in git branch test, all test will always be run.
}
\details{
Running the test suite without stan models take less than a minute,
running the basic models take roughly less than 10 minutes as of now.

Note! The testuites checks for the environment variables
TEST_STAN_BASIC_ON_LOCAL and TEST_STAN_FULL_ON_LOCAL to run the tests.
Hence this can
}
