% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_line.R
\name{time_line}
\alias{time_line}
\alias{time_line.numeric}
\alias{time_line.data.frame}
\title{Construct a time_line}
\usage{
time_line(
  x,
  time_scale,
  ...,
  week_start = getOption("lubridate.week.start", 1)
)

\method{time_line}{numeric}(
  x,
  time_scale,
  start_date,
  ...,
  week_start = getOption("lubridate.week.start", 1)
)

\method{time_line}{data.frame}(
  x,
  time_scale,
  start_date,
  ...,
  week_start = getOption("lubridate.week.start", 1)
)
}
\arguments{
\item{x}{a vector of the same length as the [time_line],
a [polls_data] object, or a [time_range] object.}

\item{time_scale}{The time_scale, [day], [week] or [month].}

\item{...}{Further arguments to methods.}

\item{week_start}{What is week starting day (1 = Monday, Default or 7 = Sunday).}

\item{start_date}{The starting date of the [time_line].}
}
\value{
a [time_line] object. A list with two data
frames, one for a [daily] [time_line] and one with the
[time_line] for the chosen [time_scale].
Both [data.frames] consist of a [date] variable and
a [t] index column.
}
\description{
Construct a timeline to use based on
the [time_scale] of interest.
}
\details{
The timeline is created based on full months, week or days.
If a [start_date] is supplied, length(x) no of days/weeks/months
units are added, including the day/week/month of [start_date].
Hence, the number of total days of the [time_line] will change,
based on the weekday of [start_date].

If [time_range] is supplied, the [time_line] will include every
day/week/month that overlaps the closed interval defined by
[time_range], including partial periods at both boundaries.
}
