% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_line.R
\name{time_scale_dates}
\alias{time_scale_dates}
\title{Convert from days to time scale}
\usage{
time_scale_dates(
  x,
  time_scale,
  week_start = getOption("lubridate.week.start", 1)
)
}
\arguments{
\item{x}{a vector of dates to convert to a specific
[time_scale]}

\item{time_scale}{to use.}

\item{week_start}{Day on which week starts following
ISO conventions - 1 means Monday (default),
7 means Sunday. You can set lubridate.week.start
option to control this parameter globally.}
}
\description{
Convert each date in [x] to the start date of its corresponding
[time_scale] bucket/latent slot. For example, dates can be mapped to the same
week start or month start when constructing a latent time line.
}
