% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{write_report}
\alias{write_report}
\title{Write reports for [poll_of_polls] objects}
\usage{
write_report(
  ...,
  report,
  parameters = NULL,
  output_file = NULL,
  output_format = "pdf_document",
  output_dir = getwd()
)
}
\arguments{
\item{...}{one or more [poll_of_polls] objects or path to rds files}

\item{report}{The report to generate (out of [supported_reports()])}

\item{parameters}{Paramaters to the individual report. See details.}

\item{output_file}{The file file name to write to}

\item{output_format}{The output format for the report}

\item{output_dir}{the output directory of the report}
}
\description{
Write reports for [poll_of_polls] objects
}
\details{
Each report has its own parameter settings.
See object param_default in the individual rmd report.
}
