% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_get_sampler_state}
\alias{backend_get_sampler_state}
\title{Backend adaptation info}
\usage{
backend_get_sampler_state(backend, fit, ...)
}
\description{
Backend adaptation info
}
\keyword{internal}
