% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_sample}
\alias{backend_sample}
\title{Run a Stan fit using the selected backend}
\usage{
backend_sample(
  backend,
  sample_arguments,
  stan_file = NULL,
  model_name = NULL,
  compile_arguments = NULL
)
}
\description{
Run a Stan fit using the selected backend
}
\keyword{internal}
