% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls_extract.R
\name{extract_pop_election_period}
\alias{extract_pop_election_period}
\title{Extract election periods from a poll_of_polls object}
\usage{
extract_pop_election_period(pop, election_period_length = 6)
}
\arguments{
\item{pop}{a poll of polls object}

\item{election_period_length}{the number of time points that make up the lection period}
}
\description{
Extract election periods from a poll_of_polls object
}
\details{
Extracts a tibble with the election period as a boolean per time point
}
