% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_get_functions.R
\name{log_prob}
\alias{log_prob}
\alias{log_prob.poll_of_polls}
\alias{log_prob.default}
\alias{get_num_upars}
\alias{get_num_upars.poll_of_polls}
\alias{get_num_upars.default}
\alias{get_num_pars}
\title{Stan Function for pop objects}
\usage{
log_prob(object, ...)

log_prob.poll_of_polls(object, ...)

log_prob.default(object, ...)

get_num_upars(object, ...)

get_num_upars.poll_of_polls(object, ...)

get_num_upars.default(object, ...)

get_num_pars(object)
}
\arguments{
\item{object}{an object}

\item{...}{arguments supplied to the rstan function}
}
\description{
Stan Function for pop objects
}
