% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_parameters_areas}
\alias{plot_parameters_areas}
\alias{plot_parameters_mcmc_hex}
\alias{plot_parameters_mcmc_scatter}
\alias{plot_parameters_mcmc_intervals}
\alias{pairs.poll_of_polls}
\alias{traceplot}
\alias{traceplot.poll_of_polls}
\title{Plot posterior distributions of parameters}
\usage{
plot_parameters_areas(
  x,
  params,
  params_plot_name = NULL,
  title = "Posterior distributions",
  ...
)

plot_parameters_mcmc_hex(
  x,
  params,
  params_plot_name = NULL,
  title = "Posterior distributions",
  ...
)

plot_parameters_mcmc_scatter(
  x,
  params,
  params_plot_name = NULL,
  title = "Posterior distributions",
  ...
)

plot_parameters_mcmc_intervals(
  x,
  params,
  params_plot_name = NULL,
  title = "Posterior distributions",
  ...
)

\method{pairs}{poll_of_polls}(x, ...)

traceplot(x, ...)

\method{traceplot}{poll_of_polls}(x, pars = NULL, inc_warmup = FALSE, ...)
}
\arguments{
\item{x}{a [poll_of_polls] object}

\item{params}{parameters to plot (see \code{parameter_names()} too see all parameters in model)}

\item{params_plot_name}{names to use for parameters in plot}

\item{title}{The title of the plot.}

\item{...}{further arguments sent to bayesplot::mcmc_areas, bayesplot::mcmc_hex, etc.}

\item{pars}{optional parameter names to include in the traceplot.}

\item{inc_warmup}{should warmup draws be included when constructing the
traceplot input?}
}
\description{
Plot posterior distributions of parameters
}
