% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_poll_of_polls}
\alias{plot_poll_of_polls}
\alias{plot.poll_of_polls}
\alias{geom_pop_line}
\title{Plot a [poll_of_polls] object}
\usage{
plot_poll_of_polls(
  x,
  y = NULL,
  from = NULL,
  to = NULL,
  publish_date = TRUE,
  collection_period = FALSE,
  shift_latent_days = 0,
  house = NULL,
  include_latent_state = TRUE,
  ...
)

\method{plot}{poll_of_polls}(x, ...)

geom_pop_line(x, y, ...)
}
\arguments{
\item{x}{a [poll_of_polls] object}

\item{y}{column to visualize.}

\item{from}{plot from [date]}

\item{to}{plot to [date]}

\item{publish_date}{visualize publication dates.}

\item{collection_period}{visualize collection period.}

\item{shift_latent_days}{Shift the latent series right, this number of days.}

\item{house}{only plot the following polling houses}

\item{include_latent_state}{include latent state in plot}

\item{...}{further arguments to [geom_line]}
}
\description{
Plot a [poll_of_polls] object
}
