% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll_of_polls.R
\name{poll_of_polls}
\alias{poll_of_polls}
\title{Fit a poll of polls object}
\usage{
poll_of_polls(
  y,
  model,
  polls_data,
  time_scale,
  known_state = NULL,
  model_time_range = NULL,
  latent_time_ranges = NULL,
  hyper_parameters = NULL,
  slow_scales = NULL,
  ...,
  cache_dir = file.path(tempdir(), "pop_cache")
)
}
\arguments{
\item{y}{character vector indicating poll variable to use}

\item{model}{poll of polls model to use (or a path to a stan model)}

\item{polls_data}{a [polls_data] object}

\item{time_scale}{the time scale to use, [day], [week], or [month].}

\item{known_state}{a [data.frame] with variables [date] and [x] indicating the known states for certain dates.}

\item{model_time_range}{a [time_range] object that describe the period used for the latent state. Default is the [time_range] of the [polls_data] object.}

\item{latent_time_ranges}{time ranges of the latent state of individual [y]s}

\item{hyper_parameters}{hyperparameters to supply direct to the model}

\item{slow_scales}{a vector of [Date]s that indicate breaks (right-inclusive) for a slower moving time scale.}

\item{...}{further arguments to [rstan::stan()] function}

\item{cache_dir}{directory to cache model. Default is cache in tempdir(). [NULL], no cache.}
}
\description{
Fit a poll of polls object
}
\details{
The [input_args] slot contain all input arguments except polls data and known state that are stored in the original object instead.
The [stan_arguments] slot contains all arguments that are supplied to the [rstan::stan()] function, except for the data argument which is stored in the [stan_data] slot.
}
