% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_data.R
\name{stan_polls_data}
\alias{stan_polls_data}
\alias{stan_polls_data_model2}
\alias{stan_polls_data_model3}
\alias{stan_polls_data_model5}
\alias{stan_polls_data_model6}
\alias{stan_polls_data_model6b}
\alias{stan_polls_data_model7}
\alias{stan_polls_data_model8a}
\alias{stan_polls_data_model8a1}
\alias{stan_polls_data_model8b}
\alias{stan_polls_data_model8c}
\alias{stan_polls_data_model8d}
\title{Generate Stan Data from [polls_data] objects}
\usage{
stan_polls_data(
  x,
  y_name,
  model,
  time_scale = "week",
  time_scale_overrides = NULL,
  known_state = NULL,
  model_time_range = NULL,
  latent_time_ranges = NULL,
  hyper_parameters = NULL,
  slow_scales = NULL
)

stan_polls_data_model2(x, y_name, time_scale = "week", model_time_range = NULL)

stan_polls_data_model3(x, y_name, time_scale = "week", model_time_range = NULL)

stan_polls_data_model5(
  x,
  y_name,
  time_scale = "week",
  known_state = NULL,
  model_time_range = NULL
)

stan_polls_data_model6(
  x,
  y_name,
  time_scale = "week",
  known_state = NULL,
  model_time_range = NULL
)

stan_polls_data_model6b(
  x,
  y_name,
  time_scale = "week",
  known_state = NULL,
  model_time_range = NULL,
  latent_time_ranges = NULL
)

stan_polls_data_model7(
  x,
  y_name,
  time_scale = "week",
  known_state = NULL,
  model_time_range = NULL,
  latent_time_ranges = NULL
)

stan_polls_data_model8a(
  x,
  y_name,
  time_scale = "week",
  known_state = NULL,
  model_time_range = NULL,
  latent_time_ranges = NULL,
  hyper_parameters = NULL
)

stan_polls_data_model8a1(
  x,
  y_name,
  time_scale = "week",
  known_state = NULL,
  model_time_range = NULL,
  latent_time_ranges = NULL,
  hyper_parameters = NULL
)

stan_polls_data_model8b(
  x,
  y_name,
  time_scale = "week",
  known_state = NULL,
  model_time_range = NULL,
  latent_time_ranges = NULL,
  hyper_parameters = NULL,
  slow_scales = NULL
)

stan_polls_data_model8c(
  x,
  y_name,
  time_scale = "week",
  known_state = NULL,
  model_time_range = NULL,
  latent_time_ranges = NULL,
  hyper_parameters = NULL,
  slow_scales = NULL
)

stan_polls_data_model8d(
  x,
  y_name,
  time_scale = "week",
  known_state = NULL,
  model_time_range = NULL,
  latent_time_ranges = NULL,
  hyper_parameters = NULL,
  slow_scales = NULL,
  model
)
}
\arguments{
\item{x}{a [polls_data] object}

\item{y_name}{a character vector indicating y variables in polls object.}

\item{model}{model to get data for.}

\item{time_scale}{to use.}

\item{time_scale_overrides}{a [data.frame] with columns [from], [to], and [time_scale]
that override the default [time_scale] for inclusive date ranges in the latent state.
The [from] and [to] bounds are inclusive and override ranges must not overlap.}

\item{known_state}{known time points in the latent state}

\item{model_time_range}{the time range to model (for example for extrapolation)}

\item{latent_time_ranges}{the time range of the latent state}

\item{hyper_parameters}{a list with hyper parameters supplied to the model.}

\item{slow_scales}{a vector of [Date]s that indicate breaks (right-inclusive) for a slower moving time scale.
Example: If only 2010-01-15 is used, all dates up to and including 2010-01-15, will have s=1,
         Dates after 2010-01-15 will have s=2.}
}
\description{
Generate Stan Data from [polls_data] objects
}
\details{
The returned object keeps the existing [stan_data] and [time_line] fields
unchanged for current models. A parallel future path is attached in
[stan_data_with_overrides] and [time_line_with_overrides], built using
[time_line_with_overrides()].
}
